solana-arbitrage-bot/requirements.txt
httpx==0.25.2
aiohttp==3.9.1
solana-py==0.32.0
structlog==23.2.0
asyncio-throttle==1.0.2
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
click==8.1.7
pytest==7.4.3
pytest-asyncio==0.21.1
base58==2.1.1